Refactor `cmap_discretisation()` and `colour_bar_index()`
